requirements.txt
Flask==2.1.2
tensorflow==2.9.1
Pillow==9.1.0
